+++ nym-wallet/src/pages/auth/pages/connect-password.tsx
@@ -2,17 +2,17 @@ import React, { useContext, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { Button, CircularProgress, FormControl, Stack } from '@mui/material';
 import { useSnackbar } from 'notistack';
+import { AuthContext } from 'src/context/auth';
+import { createPassword } from 'src/requests';
+import { PasswordInput } from 'src/components';
 import { Subtitle, Title, PasswordStrength } from '../components';
-import { PasswordInput } from '../components/textfields';
-import { SignInContext } from '../context';
-import { createPassword } from '../../../requests';
 
 export const ConnectPassword = () => {
   const [confirmedPassword, setConfirmedPassword] = useState<string>('');
   const [isStrongPassword, setIsStrongPassword] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
 
-  const { mnemonic, password, setPassword, resetState } = useContext(SignInContext);
+  const { mnemonic, password, setPassword, resetState } = useContext(AuthContext);
   const history = useHistory();
 
   const { enqueueSnackbar } = useSnackbar();