+++ nym-wallet/src/pages/bond/SuccessView.tsx
@@ -1,11 +1,11 @@
 import React, { useContext } from 'react';
 import { Box } from '@mui/material';
 import { SuccessReponse, TransactionDetails } from '../../components';
-import { ClientContext } from '../../context/main';
+import { AppContext } from '../../context/main';
 import { useCheckOwnership } from '../../hooks/useCheckOwnership';
 
 export const SuccessView: React.FC<{ details?: { amount: string; address: string } }> = ({ details }) => {
-  const { userBalance, currency } = useContext(ClientContext);
+  const { userBalance, currency } = useContext(AppContext);
   const { ownership } = useCheckOwnership();
 
   return (